Add files via upload
audioControl-effects for audio
live_play-going to experiment with midi keyboard/mixer, would be cool to get this implemented into audioControl.py for realtime audio playback. Sounds cool just as is for now though.
GUI-Havent tried that hard to get it to look like the mock up from the presentation.
SpectrumAnalyzer-Followed a tutorial online, super complicated and kind of expensive so might get rid of the idea. still needs to be implemented into the GUI